product_service/pyproject.toml
[project]
name = "product_service"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "John Doe", email = "email@example.com"},
]
dependencies = [
    "pika>=1.3.2",
    "sqlalchemy>=2.0.32",
    "psycopg2-binary>=2.9.9",
    "python-decouple>=3.8",
    "python-jose>=3.3.0",
    "python-jose[cryptography]>=3.3.0",
    "newrelic>=9.13.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
